ventilairsec2ha/rootfs/requirements.txt
pyserial==3.5
paho-mqtt==1.6.1
aiohttp==3.9.1
pyyaml==6.0.1
requests==2.31.0
cryptography==41.0.7
